refactor(RichTextEditor): extract useLinkEditor hook
Move the link-editor cluster (open-state + apply/remove/cancel handlers)
out of RichTextEditor.tsx into a dedicated useLinkEditor hook. Pure
internal refactor — no public API or behavior change. The editor keeps
the <RichTextLinkEditor> JSX render and the toolbar linkActive
derivation (both need render scope / the live selection); only the state
and handlers move. Handlers read latest doc/readOnly via stable getters
so their identity stays stable.
